x/schema split model.go into separate files
